requirements.txt
alembic==1.4.3
appnope==0.1.2
backcall==0.2.0
cffi==1.14.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.3
pyheif==0.5.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
traitlets==4.3.3
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
